add PairingNoticeList, replaces CamperList and children
The term 'camper' has been substituted for 'pairing notice' to highlight that a camper is a distinct entity that owns a notice, and a notice contains more data just the camper's details.
CamperImage replaces CamperPhoto, it encapsulates only the markup for showing the image, and does not include a link to the camper's profile, that must now be done by a parent.
PairingNotice replaces CamperCard.
PairingNoticeList replaces CamperList.
The PairingNoticeList container encapsulates all state and logic to do with showing and deleting notices.
App no longer contains any logic to do with data manipulation.
date-fns has been added to dependencies for date formatting